fix(Attemp to load infinite tiles): Fixed second layer in the stories; separated <TileLayer> tags so
