+++ packages/base-map/__mocks__/VehicleMarker.js
