+++ packages/base-map/__mocks__/vehicle-data/all-vehicles.json
@@ -1,0 +1,14 @@
+[
+  {
+    "vehicleID": 1742,
+    "routeNumber": "40",
+    "latitude": 33.755558,
+    "longitude": -84.393716
+  },
+  {
+    "vehicleID": 7024,
+    "routeNumber": "03",
+    "latitude": 33.754399,
+    "longitude": -84.37547
+  }
+]
