+++ packages/base-map/__mocks__/vehicle-data/select-vehicles.json
@@ -1,0 +1,146 @@
+[
+  {
+    "shapeId": "433108",
+    "stopSequence": 31,
+    "stopId": "10121",
+    "id": "523+537",
+    "routeShortName": "MAX Blue",
+    "vehicleId": "523+537",
+    "routeColor": null,
+    "lon": -84.399852,
+    "routeId": "100",
+    "status": "IN_TRANSIT_TO",
+    "blockId": "9024",
+    "agencyId": "TRIMET",
+    "seconds": 7,
+    "serviceId": "A.556",
+    "directionId": "1",
+    "lat": 33.749081,
+    "tripId": "9530557",
+    "routeLongName": "Blue to Hillsboro",
+    "routeType": "TRAM",
+    "reportDate": "12/06/19 8:40 AM",
+    "routeTextColor": null,
+    "heading": 234.0
+  },
+  {
+    "shapeId": "433090",
+    "stopSequence": 1,
+    "stopId": "9848",
+    "id": "210+318",
+    "routeShortName": "MAX Blue",
+    "vehicleId": "210+318",
+    "routeColor": null,
+    "lon": -84.391074,
+    "routeId": "100",
+    "status": "STOPPED_AT",
+    "blockId": "9025",
+    "agencyId": "TRIMET",
+    "seconds": 26,
+    "serviceId": "A.556",
+    "directionId": "0",
+    "lat": 33.752137,
+    "tripId": "9529799",
+    "routeLongName": "Blue to Gresham",
+    "routeType": "TRAM",
+    "reportDate": "12/06/19 8:40 AM",
+    "routeTextColor": null,
+    "heading": 180.0
+  },
+  {
+    "shapeId": "433108",
+    "stopSequence": 16,
+    "stopId": "8373",
+    "id": "241+307",
+    "routeShortName": "MAX Blue",
+    "vehicleId": "241+307",
+    "routeColor": null,
+    "lon": -84.308856,
+    "routeId": "100",
+    "status": "IN_TRANSIT_TO",
+    "blockId": "9026",
+    "agencyId": "TRIMET",
+    "seconds": 21,
+    "serviceId": "A.556",
+    "directionId": "1",
+    "lat": 33.757645,
+    "tripId": "9530561",
+    "routeLongName": "Blue to Hillsboro",
+    "routeType": "TRAM",
+    "reportDate": "12/06/19 8:40 AM",
+    "routeTextColor": null,
+    "heading": 278.0
+  },
+  {
+    "shapeId": "433108",
+    "stopSequence": 2,
+    "stopId": "8360",
+    "id": "238+303",
+    "routeShortName": "MAX Blue",
+    "vehicleId": "238+303",
+    "routeColor": null,
+    "lon": -84.126933,
+    "routeId": "100",
+    "status": "STOPPED_AT",
+    "blockId": "9027",
+    "agencyId": "TRIMET",
+    "seconds": 14,
+    "serviceId": "A.556",
+    "directionId": "1",
+    "lat": 33.732605,
+    "tripId": "9530564",
+    "routeLongName": "Blue to Hillsboro",
+    "routeType": "TRAM",
+    "reportDate": "12/06/19 8:40 AM",
+    "routeTextColor": null,
+    "heading": 285.0
+  },
+  {
+    "shapeId": "433113",
+    "stopSequence": 18,
+    "stopId": "10117",
+    "id": "243+251",
+    "routeShortName": "MAX Blue",
+    "vehicleId": "243+251",
+    "routeColor": null,
+    "lon": -84.391895,
+    "routeId": "100",
+    "status": "IN_TRANSIT_TO",
+    "blockId": "9028",
+    "agencyId": "TRIMET",
+    "seconds": 38,
+    "serviceId": "A.556",
+    "directionId": "1",
+    "lat": 33.749573,
+    "tripId": "9530558",
+    "routeLongName": "Blue to SW 170th",
+    "routeType": "TRAM",
+    "reportDate": "12/06/19 8:40 AM",
+    "routeTextColor": null,
+    "heading": 204.0
+  },
+  {
+    "shapeId": "433090",
+    "stopSequence": 44,
+    "stopId": "13450",
+    "id": "117+244",
+    "routeShortName": "MAX Blue",
+    "vehicleId": "117+244",
+    "routeColor": null,
+    "lon": -84.141048,
+    "routeId": "100",
+    "status": "STOPPED_AT",
+    "blockId": "9030",
+    "agencyId": "TRIMET",
+    "seconds": 8,
+    "serviceId": "A.556",
+    "directionId": "0",
+    "lat": 33.737893,
+    "tripId": "9529787",
+    "routeLongName": "Blue to Gresham",
+    "routeType": "TRAM",
+    "reportDate": "12/06/19 8:40 AM",
+    "routeTextColor": null,
+    "heading": 120.0
+  }
+]
